Refactor: create ccache_core lib
