+++ src/Manifest.hpp
@@ -18,7 +18,7 @@
 
 #pragma once
 
-#include "system.hpp"
+#include "core/system.hpp"
 
 #include "third_party/nonstd/optional.hpp"
