+++ src/SignalHandler.hpp
@@ -18,7 +18,7 @@
 
 #pragma once
 
-#include "system.hpp"
+#include "core/system.hpp"
 
 #include "signal.h"
