+++ src/ZstdCompressor.hpp
@@ -18,7 +18,7 @@
 
 #pragma once
 
-#include "system.hpp"
+#include "core/system.hpp"
 
 #include "Compressor.hpp"
 #include "NonCopyable.hpp"
