+++ src/ccache.hpp
@@ -19,7 +19,7 @@
 
 #pragma once
 
-#include "system.hpp"
+#include "core/system.hpp"
 
 #include "Config.hpp"
