+++ src/core/CMakeLists.txt
@@ -1,0 +1,3 @@
+add_library(ccache_core INTERFACE)
+
+target_include_directories(ccache_core INTERFACE include)
